second commit

+++ controllers/brand.controller.js
@@ -1,20 +1,28 @@
 const BrandModel = require('../models/brand.model')
-const url = require('url')
+const WatchModel = require("../models/watch.model")
 
 module.exports = {
     createBrand: async (req, res) => {
         try {
             const bodyData = req.body
             if(bodyData.brandName == ""){
-                return res.render('admin/brand/create-brand.ejs', {errorMessage: "Cannot be left blank",  layout: "admin/masterDashboard.ejs"})
+                return res.render('admin/brand/create-brand.ejs', {
+                    brand: bodyData,
+                    errorMessage: "Cannot be left blank",  
+                    layout: "admin/masterDashboard.ejs"
+                })
             }
 
             const isBrandNameExited = await BrandModel.findOne({brandName: bodyData.brandName})
             if (isBrandNameExited) {
-                return res.render('admin/brand/create-brand.ejs', {errorMessage: "Must not be the same as an existing name",  layout: "admin/masterDashboard.ejs"})
+                return res.render('admin/brand/create-brand.ejs', {
+                    brand: bodyData,
+                    errorMessage: "Must not be the same as an existing name", 
+                    layout: "admin/masterDashboard.ejs"
+                })
             }
             
-            const newBrand = await BrandModel.create(bodyData)
+            await BrandModel.create(bodyData)
 
             // return res.render ("admin/brand/create-brand.ejs", {errorMessage: "Create successfully", layout: "admin/masterDashboard.ejs"})
             res.redirect("/brands")
@@ -23,13 +31,26 @@ module.exports = {
         }
     },
     renderCreateBrandPage: async (req, res) => {
-        res.render("admin/brand/create-brand.ejs", { errorMessage: "", layout: "admin/masterDashboard.ejs"})
+        const bodyData = req.body
+        try {
+            res.render("admin/brand/create-brand.ejs", { 
+                brand: bodyData,
+                errorMessage: "", 
+                layout: "admin/masterDashboard.ejs"
+            })
+        } catch (error) {
+            console.log(error.message)
+        }
     },
     getBrand: async (req, res) => {
         try {
             const response = await BrandModel.find()
             // console.log(response)
-            return res.render("admin/brand/get-brand.ejs", {brands: response, layout: "admin/masterDashboard.ejs"});
+            return res.render("admin/brand/get-brand.ejs", {
+                brands: response, 
+                errorMessage: "", 
+                layout: "admin/masterDashboard.ejs"
+            });
         } catch (error) {
             console.log(error.message)
         }
@@ -38,6 +59,16 @@ module.exports = {
         const brandId = req.params.id // trong đg link là /deleteProduct/:id nên nhớ viết id thoy
         // console.log(productId)
         try {
+            const response = await BrandModel.find()
+            const exitedBrandOfWatch = await WatchModel.findOne({brand: brandId})
+            // console.log(exitedBrandOfWatch)
+            if (exitedBrandOfWatch){
+                return res.render("admin/brand/get-brand.ejs", {
+                    brands: response, 
+                    errorMessage: "The brand cannot be deleted because there is still a watch for this brand", 
+                    layout: "admin/masterDashboard.ejs"});
+                // res.send("The brand cannot be deleted because there is still a watch for this brand")
+            }
             await BrandModel.findOneAndDelete({_id: brandId})
             res.redirect("/brands")
         } catch (error) {
@@ -48,7 +79,11 @@ module.exports = {
         const brandId = req.params.id
         try {
             const response = await BrandModel.findOne({_id: brandId})
-            res.render("admin/brand/edit-brand", {errorMessage: "", brand: response, layout: "admin/masterDashboard.ejs"})
+            res.render("admin/brand/edit-brand", {
+                errorMessage: "", 
+                brand: response, 
+                layout: "admin/masterDashboard.ejs"
+            })
         } catch (error) {
             console.log(error.message)
         }
@@ -61,7 +96,18 @@ module.exports = {
             const response = await BrandModel.findOne({_id: brandId})
             const isBrandNameExited = await BrandModel.findOne({brandName: bodyData.brandName})
             if (isBrandNameExited) {
-                return res.render('admin/brand/edit-brand.ejs', {errorMessage: "Please update the new name", brand: response, layout: "admin/masterDashboard.ejs"})
+                return res.render('admin/brand/edit-brand.ejs', {
+                    errorMessage: "Please update the new name", 
+                    brand: response, 
+                    layout: "admin/masterDashboard.ejs"
+                })
+            }
+            if(bodyData.brandName == ""){
+                return res.render('admin/brand/edit-brand.ejs', {
+                    errorMessage: "Cannot be left blank", 
+                    brand: response, 
+                    layout: "admin/masterDashboard.ejs"
+                })
             }
 
             await BrandModel.findByIdAndUpdate(brandId, bodyData)
